fix(operation_ui): guard description JSON.parse in triage & project fields
Addresses CodeRabbit review on PR #7907: the unguarded JSON.parse(description)
render crash fixed in TaskFields also existed in TriageFields and ProjectFields
(plain-text / legacy descriptions would throw during render and crash the page).

Extract the defensive parse into a shared parseDescriptionBlocks() helper and use
it across all three components, removing the duplicated inline logic.
