+++ frontend/plugins/operation_ui/src/modules/task/components/detail/TaskFields.tsx
@@ -23,24 +23,7 @@ import { useEffect, useRef, useState } from 'react';
 import { useDebounce } from 'use-debounce';
 import { TagsSelect } from 'ui-modules';
 import { IconTags } from '@tabler/icons-react';
-
-// Task descriptions are normally BlockNote JSON (a Block[]), but legacy or
-// API/script-created tasks can hold plain text. Parsing defensively keeps a
-// non-JSON description from throwing during render and crashing the page —
-// the text is rendered as a single paragraph instead.
-const parseDescriptionBlocks = (
-  description?: string,
-): Block[] | undefined => {
-  if (!description) return undefined;
-  try {
-    const parsed = JSON.parse(description);
-    return Array.isArray(parsed) && parsed.length > 0
-      ? (parsed as Block[])
-      : undefined;
-  } catch {
-    return [{ type: 'paragraph', content: description } as unknown as Block];
-  }
-};
+import { parseDescriptionBlocks } from '@/operation/utils/parseDescriptionBlocks';
 
 export const TaskFields = ({ task }: { task: ITask }) => {
   const {
